Add the ability to disable user tracking. (#794, DEVEX-242)
* Added `setTrackingDisabled` and `trackingDisabled` methods to the Branch class to set and query the tracking state.

* Added the `BNCTrackingDisabledError` error code that's returned to callbacks when tracking is disabled.

* Added a `clearTrackingInformation` method to the preference helper class that clears user identifiable information from preferences.

* Changed some NSDictionaries to NSMutableDictionaries for crash safety: Storing nil to a mutable dictionary is not a crash.

* Added a 'Do not track' option to UITestBed.

* Fixed some code analyzer warnings.

* Updated the SDK so that deep links can be opened even when 'no tracking' is set.

* Added open stress test.

* Updated documentation.

* Added SMS as a link scraper when sharing.

* Fixed some unit tests and updated pods.

+++ Examples/WebViewExample/WebViewExample/Utilities/BNCLog.swift
@@ -16,7 +16,7 @@ import Branch
  *   - line: (unused) provides the Swift line number
  */
 func BNCLogDebug(_ message: String, _ file: String=#file, _ line: UInt=#line) {
-    BNCLogWriteMessage(.debug, file, line, message)
+    BNCLogWriteMessage(.debug, file, Int32(line), message)
 }
 
 /**
@@ -27,7 +27,7 @@ func BNCLogDebug(_ message: String, _ file: String=#file, _ line: UInt=#line) {
  *   - line: (unused) provides the Swift line number
  */
 func BNCLogWarning(_ message: String, _ file: String=#file, _ line: UInt=#line) {
-    BNCLogWriteMessage(.warning, file, line, message)
+    BNCLogWriteMessage(.warning, file, Int32(line), message)
 }
 
 /**
@@ -38,7 +38,7 @@ func BNCLogWarning(_ message: String, _ file: String=#file, _ line: UInt=#line) 
  *   - line: (unused) provides the Swift line number
  */
 func BNCLogError(_ message: String, _ file: String=#file, _ line: UInt=#line) {
-    BNCLogWriteMessage(.error, file, line, message)
+    BNCLogWriteMessage(.error, file, Int32(line), message)
 }
 
 /**
@@ -49,5 +49,5 @@ func BNCLogError(_ message: String, _ file: String=#file, _ line: UInt=#line) {
  *   - line: (unused) provides the Swift line number
  */
 func BNCLog(_ message: String, _ file: String=#file, _ line: UInt=#line) {
-    BNCLogWriteMessage(.log, file, line, message)
+    BNCLogWriteMessage(.log, file, Int32(line), message)
 }
